[NFC] Refactor transform dialect matcher. (#11211)
Move the matcher closer to the transform dialect creation to bypass the
attributes. Work in progress.
