+++ compiler/src/iree/compiler/Codegen/LLVMGPU/CMakeLists.txt
@@ -15,6 +15,7 @@ iree_cc_library(
     LLVMGPU
   HDRS
     "ConvertToLLVM.h"
+    "GPUTransformDialectStrategies.h"
     "KernelConfig.h"
     "TilingUtils.h"
     "TransposeUtils.h"
@@ -22,6 +23,7 @@ iree_cc_library(
     "ConvertToLLVM.cpp"
     "ConvertToNVVM.cpp"
     "ConvertToROCDL.cpp"
+    "GPUTransformDialectStrategies.cpp"
     "KernelConfig.cpp"
     "LLVMGPUDistribute.cpp"
     "LLVMGPULowerExecutableTarget.cpp"
@@ -88,6 +90,7 @@ iree_cc_library(
     MLIRVectorToSCF
     MLIRVectorTransforms
     iree::compiler::Codegen::Common
+    iree::compiler::Codegen::Common::TransformExtensions::CommonExtensions
     iree::compiler::Codegen::Dialect::IREECodegenDialect
     iree::compiler::Codegen::LLVMGPU::TransformExtensions::LLVMGPUExtensions
     iree::compiler::Codegen::PassHeaders
